// src/type/enum/national-excellent-exam.enum.ts
export enum NationalExcellentExamType {
    NATIONAL = "Học sinh giỏi cấp Quốc Gia",
}

// Define the enum for the National Excellent Student Exam subjects
export enum NationalExcellentStudentExamSubject {
    BIOLOGY = "Sinh Học",
    CHEMISTRY = "Hoá Học",
    CHINESE = "Tiếng Trung",
    ENGLISH = "Tiếng Anh",
    FRENCH = "Tiếng Pháp",
    GEOGRAPHY = "Địa Lý",
    HISTORY = "Lịch Sử",
    INFORMATION_TECHNOLOGY = "Tin Học",
    JAPANESE = "Tiếng Nhật",
    LITERATURE = "Ngữ Văn",
    MATHEMATICS = "Toán",
    PHYSICS = "Vật Lý",
    RUSSIAN = "Tiếng Nga",
}